+++ forms-flow-rsbcservice/src/component/RSBCImage.settingsForm.js
