+++ forms-flow-rsbcservice/src/component/printHelperServices.js
@@ -1,13 +1,13 @@
 import moment from "moment-timezone";
-import twentyFourHourDriverform from "./assets/MV2634E_082023_driver.png";
-import twentyFourHourILOform from "./assets/MV2634E_082023_ilo.png";
-import twentyFourHourPoliceform from "./assets/MV2634E_082023_icbc.png";
-import viDriverForm from "./assets/MV2721_202404.png";
-import viIncidentDetails from "./assets/MV2722_202404_Incident_Details.png";
-import appealsForm from "./assets/MV2721_202404_appeal.png";
-import viReportForm from "./assets/MV2722_202404.png";
-import twelveHourDriverForm from "./assets/MV2906E_082023_driver.png";
-import twelveHourICBCForm from "./assets/MV2906E_082023_icbc.png";
+import twentyFourHourDriverform from "../assets/MV2634E_082023_driver.png";
+import twentyFourHourILOform from "../assets/MV2634E_082023_ilo.png";
+import twentyFourHourPoliceform from "../assets/MV2634E_082023_icbc.png";
+import viDriverForm from "../assets/MV2721_202404.png";
+import viIncidentDetails from "../assets/MV2722_202404_Incident_Details.png";
+import appealsForm from "../assets/MV2721_202404_appeal.png";
+import viReportForm from "../assets/MV2722_202404.png";
+import twelveHourDriverForm from "../assets/MV2906E_082023_driver.png";
+import twelveHourICBCForm from "../assets/MV2906E_082023_icbc.png";
